pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-transfer"
version = "1.0.0"
description = "Transfer Claude Code agents between systems with interactive selection"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "PromptChain", email = "dev@promptchain.ai"}
]
keywords = ["claude", "claude-code", "agents", "transfer", "backup"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "markdown>=3.5.0",
    "pygments>=2.16.0",
    "jinja2>=3.1.0",
]

[project.scripts]
agent-transfer = "agent_transfer.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/promptchain/agent-transfer"
Documentation = "https://github.com/promptchain/agent-transfer#readme"
Repository = "https://github.com/promptchain/agent-transfer"

[tool.setuptools]
packages = ["agent_transfer"]

[tool.setuptools.package-data]
agent_transfer = ["*.md"]

